+++ mne/inverse_sparse/tests/test_gamma_map.py
@@ -60,7 +60,7 @@ def test_gamma_map():
     evoked = read_evokeds(fname_evoked, condition=0, baseline=(None, 0),
                           proj=False)
     evoked.resample(50, npad=100)
-    evoked.crop(tmin=0.1, tmax=0.16)  # crop to nice window near samp border
+    evoked.crop(tmin=0.1, tmax=0.16)  # crop to window around peak
 
     cov = read_cov(fname_cov)
     cov = regularize(cov, evoked.info)
@@ -95,7 +95,7 @@ def test_gamma_map_vol_sphere():
     evoked = read_evokeds(fname_evoked, condition=0, baseline=(None, 0),
                           proj=False)
     evoked.resample(50, npad=100)
-    evoked.crop(tmin=0.1, tmax=0.16)  # crop to nice window near samp border
+    evoked.crop(tmin=0.1, tmax=0.16)  # crop to window around peak
 
     cov = read_cov(fname_cov)
     cov = regularize(cov, evoked.info)